+++ tests/testthat/test_dmta.R
@@ -32,20 +32,6 @@ test_that("Different backends get consistent results for DFOP tc, dimethenamid d
   saem_saemix_dfop_tc_mkin <- saem(dmta_dfop_tc, transformations = "mkin")
   ints_saemix_mkin <- intervals(saem_saemix_dfop_tc_mkin)
 
-  # nlmixr saem
-  saem_nlmixr_dfop_tc <- nlmixr(dmta_dfop_tc, est = "saem",
-    control = nlmixr::saemControl(nBurn = 300, nEm = 100, nmc = 9, print = 0))
-  ints_nlmixr_saem <- intervals(saem_nlmixr_dfop_tc)
-
-  # nlmixr focei
-  # We get three warnings about nudged etas, the initial optimization and
-  # gradient problems with initial estimate and covariance
-  # We need to capture output, otherwise it pops up in testthat output
-  expect_warning(tmp <- capture_output(focei_nlmixr_dfop_tc <- nlmixr(
-      dmta_dfop_tc, est = "focei",
-      control = nlmixr::foceiControl(print = 0), all = TRUE)))
-  ints_nlmixr_focei <- intervals(focei_nlmixr_dfop_tc)
-
   # Fixed effects
   ## saemix vs. nlme
   expect_true(all(ints_saemix$fixed[, "est."] >
@@ -59,18 +45,6 @@ test_that("Different backends get consistent results for DFOP tc, dimethenamid d
   expect_true(all(ints_saemix_mkin$fixed[, "est."] <
       backtransform_odeparms(ints_nlme$fixed[, "upper"], dmta_dfop$mkinmod)))
 
-  ## nlmixr saem vs. nlme
-  expect_true(all(ints_nlmixr_saem$fixed[, "est."] >
-      backtransform_odeparms(ints_nlme$fixed[, "lower"], dmta_dfop$mkinmod)))
-  expect_true(all(ints_nlmixr_saem$fixed[, "est."] <
-      backtransform_odeparms(ints_nlme$fixed[, "upper"], dmta_dfop$mkinmod)))
-
-  ## nlmixr focei vs. nlme
-  expect_true(all(ints_nlmixr_focei$fixed[, "est."] >
-      backtransform_odeparms(ints_nlme$fixed[, "lower"], dmta_dfop$mkinmod)))
-  expect_true(all(ints_nlmixr_focei$fixed[, "est."] <
-      backtransform_odeparms(ints_nlme$fixed[, "upper"], dmta_dfop$mkinmod)))
-
   # Random effects
   ## for saemix with saemix transformations, the comparison would be complicated...
   ## saemix mkin vs. nlme
@@ -78,18 +52,6 @@ test_that("Different backends get consistent results for DFOP tc, dimethenamid d
       backtransform_odeparms(ints_nlme$reStruct$ds[, "lower"], dmta_dfop$mkinmod)))
   expect_true(all(ints_saemix$fixed[, "est."] <
       backtransform_odeparms(ints_nlme$fixed[, "upper"], dmta_dfop$mkinmod)))
-
-  ## nlmixr saem vs. nlme
-  expect_true(all(ints_nlmixr_saem$random[, "est."] >
-      backtransform_odeparms(ints_nlme$reStruct$ds[, "lower"], dmta_dfop$mkinmod)))
-  expect_true(all(ints_nlmixr_saem$random[, "est."] <
-      backtransform_odeparms(ints_nlme$reStruct$ds[, "upper"], dmta_dfop$mkinmod)))
-
-  ## nlmixr focei vs. nlme
-  expect_true(all(ints_nlmixr_focei$random[, "est."] >
-      backtransform_odeparms(ints_nlme$reStruct$ds[, "lower"], dmta_dfop$mkinmod)))
-  expect_true(all(ints_nlmixr_focei$random[, "est."] <
-      backtransform_odeparms(ints_nlme$reStruct$ds[, "upper"], dmta_dfop$mkinmod)))
 
   # Variance function
   # Some of these tests on error model parameters fail on Travis and Winbuilder
@@ -106,21 +68,6 @@ test_that("Different backends get consistent results for DFOP tc, dimethenamid d
       ints_nlme$varStruct[, "lower"]))
   expect_true(all(ints_saemix_mkin[[3]][, "est."] <
       ints_nlme$varStruct[, "upper"]))
-
-  # nlmixr saem vs. nlme
-  expect_true(all(ints_nlmixr_saem[[3]][, "est."] >
-      ints_nlme$varStruct[, "lower"]))
-  expect_true(all(ints_nlmixr_saem[[3]][, "est."] <
-      ints_nlme$varStruct[, "upper"]))
-
-  # nlmixr focei vs. nlme
-  # We only test for the proportional part (rsd_high), as the
-  # constant part (sigma_low) obtained with nlmixr/FOCEI is below the lower
-  # bound of the confidence interval obtained with nlme
-  expect_true(ints_nlmixr_focei[[3]]["rsd_high", "est."] >
-      ints_nlme$varStruct["prop", "lower"])
-  expect_true(ints_nlmixr_focei[[3]]["rsd_high", "est."] <
-      ints_nlme$varStruct["prop", "upper"])
 })
 
 # Compared to the 2020 paper https://doi.org/10.3390/environments8080071
@@ -147,19 +94,6 @@ test_that("Different backends get consistent results for SFO-SFO3+, dimethenamid
     start = mean_degparms(dmta_sfo_sfo3p_tc, test_log_parms = TRUE)),
     "Iteration 5, LME step.*not converge")
   ints_nlme_mets <- intervals(nlme_sfo_sfo3p_tc, which = "fixed")
-
-  # The saem fit with nlmixr takes only about 15 seconds
-  tmp <- capture.output(
-    saem_nlmixr_sfo_sfo3p_tc <- nlmixr(dmta_sfo_sfo3p_tc, est = "saem",
-      control = nlmixr::saemControl(print = 0)))
-  ints_nlmixr_saem_mets <- intervals(saem_nlmixr_sfo_sfo3p_tc)
-
-  # We need to exclude the ilr transformed formation fractions in these
-  # tests, as they do not have a one to one relation in the transformations
-  expect_true(all(ints_nlmixr_saem_mets$fixed[, "est."][-c(6, 7, 8)] >
-      backtransform_odeparms(ints_nlme_mets$fixed[, "lower"][-c(6, 7, 8)], sfo_sfo3p)))
-  expect_true(all(ints_nlmixr_saem_mets$fixed[, "est."][-c(6, 7, 8)] <
-      backtransform_odeparms(ints_nlme_mets$fixed[, "upper"], sfo_sfo3p)[-c(6, 7, 8)]))
 
   skip("Fitting this ODE model with saemix takes about 15 minutes on my system")
   # As DFOP is overparameterised and leads to instabilities and errors, we