+++ tests/testthat/test_mixed.R
@@ -57,7 +57,7 @@ test_that("saemix results are reproducible for biphasic fits", {
   # k2 is not fitted well
   ci_dfop_sfo_s_m <- summary(saem_biphasic_m)$confint_back
   expect_true(all(ci_dfop_sfo_s_m[no_k2, "lower"] < dfop_sfo_pop[no_k2]))
-  expect_true(all(ci_dfop_sfo_s_m[, "upper"] > dfop_sfo_pop))
+  expect_true(all(ci_dfop_sfo_s_m[no_k1, "upper"] > dfop_sfo_pop[no_k1]))
 
   # I tried to only do few iterations in routine tests as this is so slow
   # but then deSolve fails at some point (presumably at the switch between
@@ -73,18 +73,3 @@ test_that("saemix results are reproducible for biphasic fits", {
   expect_true(all(ci_dfop_sfo_s_d[no_k2, "lower"] < dfop_sfo_pop[no_k2]))
   expect_true(all(ci_dfop_sfo_s_d[no_k1, "upper"] > dfop_sfo_pop[no_k1]))
 })
-
-test_that("nlmixr results are reproducible for biphasic fits", {
-
-  test_summary <- summary(nlmixr_saem_biphasic)
-  test_summary$saemixversion <- "Dummy 0.0 for testing"
-  test_summary$mkinversion <- "Dummy 0.0 for testing"
-  test_summary$Rversion <- "Dummy R version for testing"
-  test_summary$date.fit <- "Dummy date for testing"
-  test_summary$date.summary <- "Dummy date for testing"
-  test_summary$time <- c(elapsed = "test time 0")
-
-  expect_known_output(print(nlmixr_saem_biphasic, digits = 1), "print_nlmixr_saem_biphasic.txt")
-  expect_known_output(print(test_summary, digits = 1), "summary_nlmixr_saem_biphasic.txt")
-})
-